refactoring and pam authentication implementation
