+++ package.json
@@ -1,12 +1,13 @@
 {
     "name": "carta-controller",
-    "version": "2.0.0-dev.21.03.23",
+    "version": "2.0.0-dev.21.03.25",
     "description": "NodeJS-based server for CARTA",
     "repository": "https://github.com/CARTAvis/carta-controller",
     "homepage": "http://cartavis.github.io/",
     "main": "dist/index.js",
     "bin": {
-        "carta-controller": "dist/carta-controller"
+        "carta-controller": "dist/carta-controller",
+        "carta-kill-script": "scripts/carta_kill_script.sh"
     },
     "scripts": {
         "start": "npx ts-node src/index.ts",
@@ -36,6 +37,7 @@
         "moment": "^2.29.1",
         "mongodb": "^3.6.2",
         "ms": "^2.1.2",
+        "node-linux-pam": "0.0.1",
         "pug": "^3.0.0",
         "read": "^1.0.7",
         "tcp-port-used": "^1.0.1",
